Some decimal API refinements. Add failing IPC test for decimals
